requirements.txt
APScheduler==3.5.3
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
lxml==4.3.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
whitenoise==4.1.2
